Fix signed-integer-overflow UB in big-endian 32-bit reads (parsers)
The format parsers assemble 32-bit big-endian values a byte at a time, e.g.
`tmp_val = *midi_data++ << 24;`. The uint8_t operand promotes to int, so when
the top byte's high bit is set, `value << 24` shifts into/past the int sign
bit -- undefined behavior (UBSan: "left shift of 255 by 24 places cannot be
represented in type 'int'"). This fires on the normal MIDI/XMI/MUS/HMP/HMI
loading paths and when reading GUS patch headers; found with AFL++ + UBSan.

Cast the shifted byte to uint32_t so the shift is well-defined unsigned
arithmetic. The computed value is unchanged on two's-complement targets; only
the `<< 24` term can overflow (`<< 16` / `<< 8` on a byte stay within int), so
those are left as-is. 25 sites across f_midi, f_xmidi, f_hmi, f_hmp, gus_pat,
and mus2mid.

+++ src/gus_pat.c
@@ -784,29 +784,29 @@ struct _sample * _WM_load_gus_pat(const char *filename, int fix_release) {
         }
 
         gus_sample->loop_fraction = gus_patch[gus_ptr + 7];
-        gus_sample->data_length = (gus_patch[gus_ptr + 11] << 24)
+        gus_sample->data_length = ((uint32_t)gus_patch[gus_ptr + 11] << 24)
                                 | (gus_patch[gus_ptr + 10] << 16)
                                 | (gus_patch[gus_ptr + 9]  <<  8)
                                 |  gus_patch[gus_ptr + 8];
-        gus_sample->loop_start  = (gus_patch[gus_ptr + 15] << 24)
+        gus_sample->loop_start  = ((uint32_t)gus_patch[gus_ptr + 15] << 24)
                                 | (gus_patch[gus_ptr + 14] << 16)
                                 | (gus_patch[gus_ptr + 13] <<  8)
                                 |  gus_patch[gus_ptr + 12];
-        gus_sample->loop_end    = (gus_patch[gus_ptr + 19] << 24)
+        gus_sample->loop_end    = ((uint32_t)gus_patch[gus_ptr + 19] << 24)
                                 | (gus_patch[gus_ptr + 18] << 16)
                                 | (gus_patch[gus_ptr + 17] <<  8)
                                 |  gus_patch[gus_ptr + 16];
         gus_sample->rate        = (gus_patch[gus_ptr + 21] << 8)
                                 |  gus_patch[gus_ptr + 20];
-        gus_sample->freq_low    = (gus_patch[gus_ptr + 25] << 24)
+        gus_sample->freq_low    = ((uint32_t)gus_patch[gus_ptr + 25] << 24)
                                 | (gus_patch[gus_ptr + 24] << 16)
                                 | (gus_patch[gus_ptr + 23] <<  8)
                                 |  gus_patch[gus_ptr + 22];
-        gus_sample->freq_high   = (gus_patch[gus_ptr + 29] << 24)
+        gus_sample->freq_high   = ((uint32_t)gus_patch[gus_ptr + 29] << 24)
                                 | (gus_patch[gus_ptr + 28] << 16)
                                 | (gus_patch[gus_ptr + 27] <<  8)
                                 |  gus_patch[gus_ptr + 26];
-        gus_sample->freq_root   = (gus_patch[gus_ptr + 33] << 24)
+        gus_sample->freq_root   = ((uint32_t)gus_patch[gus_ptr + 33] << 24)
                                 | (gus_patch[gus_ptr + 32] << 16)
                                 | (gus_patch[gus_ptr + 31] <<  8)
                                 |  gus_patch[gus_ptr + 30];
